Expose an `update_type` helper function.
