+++ src/types/summary_helper.rs
@@ -62,46 +62,7 @@ enum BlockItemType {
     #[serde(rename = "updateTransaction")]
     /// Chain updates are signed by the governance keys. They affect the core
     /// parameters of the chain.
-    Update(UpdateType),
-}
-
-#[derive(SerdeSerialize, SerdeDeserialize, Debug, Clone, Copy)]
-#[serde(rename_all = "camelCase")]
-// Since all variants are fieldless, the default JSON serialization will convert
-// all the variants to simple strings.
-/// Enumeration of the types of updates that are possible.
-pub enum UpdateType {
-    /// Update the chain protocol
-    UpdateProtocol,
-    /// Update the election difficulty
-    UpdateElectionDifficulty,
-    /// Update the euro per energy exchange rate
-    UpdateEuroPerEnergy,
-    /// Update the microCCD per euro exchange rate
-    UpdateMicroGTUPerEuro,
-    /// Update the address of the foundation account
-    UpdateFoundationAccount,
-    /// Update the distribution of newly minted CCD
-    UpdateMintDistribution,
-    /// Update the distribution of transaction fees
-    UpdateTransactionFeeDistribution,
-    /// Update the GAS rewards
-    UpdateGASRewards,
-    /// Minimum amount to register as a baker
-    UpdateBakerStakeThreshold,
-    /// Add new anonymity revoker
-    UpdateAddAnonymityRevoker,
-    /// Add new identity provider
-    UpdateAddIdentityProvider,
-    /// Update the root keys
-    UpdateRootKeys,
-    /// Update the level 1 keys
-    UpdateLevel1Keys,
-    /// Update the level 2 keys
-    UpdateLevel2Keys,
-    UpdatePoolParameters,
-    UpdateCooldownParameters,
-    UpdateTimeParameters,
+    Update(super::UpdateType),
 }
 
 #[derive(SerdeSerialize, SerdeDeserialize, Debug, Clone)]
@@ -382,31 +343,6 @@ pub(crate) enum Event {
         /// Delegator account
         account:      AccountAddress,
     },
-}
-
-impl super::UpdatePayload {
-    fn update_type(&self) -> UpdateType {
-        use UpdateType::*;
-        match self {
-            UpdatePayload::Protocol(_) => UpdateProtocol,
-            UpdatePayload::ElectionDifficulty(_) => UpdateElectionDifficulty,
-            UpdatePayload::EuroPerEnergy(_) => UpdateEuroPerEnergy,
-            UpdatePayload::MicroGTUPerEuro(_) => UpdateMicroGTUPerEuro,
-            UpdatePayload::FoundationAccount(_) => UpdateFoundationAccount,
-            UpdatePayload::MintDistribution(_) => UpdateMintDistribution,
-            UpdatePayload::TransactionFeeDistribution(_) => UpdateTransactionFeeDistribution,
-            UpdatePayload::GASRewards(_) => UpdateGASRewards,
-            UpdatePayload::BakerStakeThreshold(_) => UpdateBakerStakeThreshold,
-            UpdatePayload::Root(_) => UpdateRootKeys,
-            UpdatePayload::Level1(_) => UpdateLevel1Keys,
-            UpdatePayload::AddAnonymityRevoker(_) => UpdateAddAnonymityRevoker,
-            UpdatePayload::AddIdentityProvider(_) => UpdateAddIdentityProvider,
-            UpdatePayload::CooldownParametersCPV1(_) => UpdateCooldownParameters,
-            UpdatePayload::PoolParametersCPV1(_) => UpdatePoolParameters,
-            UpdatePayload::TimeParametersCPV1(_) => UpdateTimeParameters,
-            UpdatePayload::MintDistributionCPV1(_) => UpdateMintDistribution,
-        }
-    }
 }
 
 impl From<super::ContractTraceElement> for Event {
